Issue #8: no modules on root (#10)
* feat: refactoring into SmartObjectFramework Module

it is now possible to install a module through a Solidity script containing two lines:

```
SmartObjectFrameworkModule module = new SmartObjectFrameworkModule();
IBaseWorld(worldAddress).installModule(module, abi.encode(namespace));
```

`Deploying with namespace==bytes14(0)` is forbidden for security reasons;
the fact that _namespace() has been added to `EveSystem` means it will probably break the current Access Control PR since there's systems there that also have their own implementation of it, but since the PR has not been merged yet it's okay.

* feat: adding SmartObjectLib

There's a misconfig somewhere and it's driving me insane
idk if it's, like, a wrong tableID or some registration tomfoolery at play, or just some general deployment shennanigan but there's *something* happening with the Test smart-deployable-system

Anyways, I know the structure is (for the most part,) fine, so it's just a matter of debugging this.

* fix: namespacing issue

This is a dirty fix, where we need to manually make sure that the namespace used to access tables in EveSystem matches the one given during the Smart Object Framework deployment;
because we don't know in advance where that module is going to be deployed exactly, and because the goal is for contracts to inherit `EveSystem` methods everywhere, we can't just use whatever _namespace() returns because it would only work for core contracts and nothing else

e.g. builders trying to deploy new systems, _namespace() method inside EveSystem will fetch the namespace they deployed their stuff to, and that's bad

anyways. There might be a clean and efficient way to do this that doesnt require to use hard-coded values, but I don;'t see it for now. It's not meant for us to deploy a bunch of core contract suites anywaym it only needs to be done once, so hardcoding stuff is fine and gas-efficient, at the cost of potentially breaking stuff if we need to re-deploy a new suite in the future. If anything, it shouldnt be too hard to catch that

* feat: refactor smart-character

same as smart-object-framework,

added unit tests

* chore: linting

* Update core/smart-object-framework/src/SmartObjectFrameworkModule.sol

Co-authored-by: skygirl <100742846+0xxlegolas@users.noreply.github.com>

* Update SmartObjectFrameworkModule.sol

* remove storeArgument flag from mud.config.ts

as 0xxlegolas said it's defaulted to true so there's no need to explicitly add this flag

* move Interfaces in SmartObjectLib to standalone files

* TODO: brittle Core Interfaces' dependancy

* Revert "feat: refactor smart-character"

This reverts commit ffbea1806274318023e31b5c6df5e23c19bd17e6.

* Update EveSystem.sol

* Update .gitignore

temporarily adding all `worlds.json` files to `.gitignore` since it's cluttering up commits at increasing speeds

---------

Co-authored-by: skygirl <100742846+0xxlegolas@users.noreply.github.com>