improvement: SOF v2 access config update for CallAccess configuration (#456)
This PR is for allowing post deployment configuration of the CallAccess
map by the deployer account, and then using this table to allow
pre-configured access for various base system-calling in the SOF. If we
like this pattern we may adopt it further later as a standard pattern in
our access control logic configurations.

- `CallAccess` Table added to the sof-v2 `mud.config`
-  `CallAccess` checking added to `AccessconfigSystem.sol` functions
- Update all hardcoded `SystemId` based access rules in
`SOFAccessSystem` to use `CallAccess` defined values instead of
hardcoded values  
- fixed the order of operations in `deleteClass` to remove data from
various places in the correct order
- Adjust `SOFAccessSystem._getClassId` to return `0` if `entityId` is
`0`(instead of throwing a silent error)  
- Add Scripts for `SetSOFCallAccess`, `EntitySystemAccessConfig`,
`TagSystemAccessConfig`, and `RoleManagementSystemAccessConfig` 
- Updateed unit tests to use the same configuration logic as in the
scripts and test the changes.